+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Do not raise KeyError if user is not found, raise Unauthorized
+  [vangheem]
 
 
 1.0.0 (2017-04-24)